Changes reguarding stories plus more
Basic stories is implemented with these changes
Recycler View for stories is now clickable with a sample of how they woud work and look like.
Picasso is added as a dependency for online image loading
network_security_config.xml is added for proper internet functioning in newer API
Grouped ActivityMain into a nested scroll view so that it works with screen of smaller sizes
same is to be done with the login screen.